+++ cli/show.c
@@ -93,7 +93,7 @@ static int show_meta(const char* target) {
             char buf[256];
             rawstd_bytes_to_size(meta->spec.size, buf, sizeof(buf));
             printf("  size: %s\n", buf);
-            printf("  mirrors: %u\n", meta->spec.mirrors);
+            printf("  mirrors: %u\n", meta->spec.width);
             printf(
                 "  state: %s\n", sync_state_to_string(meta->sync_state.state)
             );
@@ -135,7 +135,7 @@ int rawstor_cli_show(const char* target, int verbose) {
 
     printf("target: %s\n", target);
     printf("size: %s\n", buf);
-    printf("mirrors: %u\n", spec.mirrors);
+    printf("mirrors: %u\n", spec.width);
 
     if (!verbose) {
         return EXIT_SUCCESS;