+++ pyrawstor/tests/test_location.py
@@ -9,7 +9,7 @@ class TestLocation(unittest.TestCase):
         with tempfile.TemporaryDirectory() as temp_dir:
             location = rawstor.Location(f"file://{temp_dir}")
 
-            target = location.create(size=1 << 20, mirrors=1)
+            target = location.create(size=1 << 20, width=1)
 
             read_targets = list(location)
             self.assertEqual(read_targets, [target])
@@ -28,7 +28,7 @@ class TestLocation(unittest.TestCase):
 
             targets = []
             for _ in range(10):
-                target = location.create(size=1 << 20, mirrors=1)
+                target = location.create(size=1 << 20, width=1)
                 targets.append(target)
             targets.sort()
 
@@ -56,7 +56,7 @@ class TestLocation(unittest.TestCase):
             self.assertEqual(info.used, 0)
             self.assertGreater(info.total, 0)
 
-            target = location.create(size=1 << 20, mirrors=1)
+            target = location.create(size=1 << 20, width=1)
 
             info = location.info()
             self.assertEqual(info.used, (1 << 20) + META_FILE_SIZE)