+++ src/backend.cpp
@@ -27,10 +27,10 @@ Backend::Backend(Private, rawio::Queue& queue, const rawstd::URI& location) :
 }
 
 void Backend::_validate_spec(const RawstorObjectSpec& sp) {
-    if (sp.mirrors != 1) {
+    if (sp.width != 1) {
         rawstd_error(
-            "Backend can only create objects with mirrors == 1, got %u\n",
-            sp.mirrors
+            "Backend can only create objects with width == 1, got %u\n",
+            sp.width
         );
         RAWSTD_THROW_SYSTEM_ERROR(EINVAL);
     }