+++ tests/test_blk_backend.cpp
@@ -100,7 +100,7 @@ rawstor::blk::Backend* open_blk_backend(
 
     rawstor::Target target({rawstd::URI(location, uuid_string)});
 
-    RawstorObjectSpec spec{.size = 1u << 20, .mirrors = 1};
+    RawstorObjectSpec spec{.size = 1u << 20, .width = 1};
     run(queue, target.create(queue, spec));
 
     object = run(queue, rawstor::Chunk::create(queue, target));