+++ tests/test_chunk.cpp
@@ -92,7 +92,7 @@ open_object(rawio::Queue& queue, const rawstd::URI& location) {
 
     rawstor::Target target({rawstd::URI(location, uuid_string)});
 
-    RawstorObjectSpec spec{.size = 1u << 20, .mirrors = 1};
+    RawstorObjectSpec spec{.size = 1u << 20, .width = 1};
     run(queue, target.create(queue, spec));
 
     return run(queue, rawstor::Chunk::create(queue, target));